add compilation to prior predictive sampler
We add a function that takes a function as an input and return a
JIT-compiled function that returns prior predictive samples from
a model definition.

The compiler relies on Python's `ast` module to parse the source code of
the function and transform random variable assignments to assignment to
samples from the distribution.