add wrapper to enforce the support
Distributions are defined on a support, but sometimes (as is the case
with the Bernoulli definition) the mathematical expression for the
logpdf will output a value even when the input is out of support (say a
float for Bernoulli).

In order to enforce the support, we borrow an idea from the Numpyro
project and add a decorator that checks that the input to the logpdf
belongs to the support.