constraint "interval" now means open interval
The "interval" constraint includes the upper and lower bounds. However,
as is the case with the Beta Distribution, they often need to be
excluded. We thus modify the `_Interval` constraint to exclude the lower
and upper bound.

If it later turns out that in some cases we need to include the upper
and lower bound we will create a separate constraint for open and closed
intervals.