Add the model class
We add the class that allows to transform a model definition function
into an actual probabilistic model from which one can get forward or
posterior samples.

It is not yet clear how we should make the distinction between a model
that takes parameters and one that takes data as an input. Maybe all
parameters & data should be considered on an equal footing, and we
should add more magic so that when `x @ mymodel(a, b)` is called we
modify the code to call `model(mymodel)` before adding the `sample` or
`logpdf` method. We could also add a `@distibution` decorator that
signals that we want to use this as a distribution and not a generative
model.