Add forward sampler
Unlike the prior sampler, the forward sampler only returns a sample
of the returned value of a generative model. This construct is not
only needed for prior predictive sampling, but also when models
are used as distributions.

In the example `a @ Horseshoe()` where `Horseshoe` is a prior written as
a probabilistic model, we are only interested by the value of the
returned value, and not the auxiliary variables. Since `sample` means
"sampling from all distributions in the model" we introduce a different
`forward` method.

We also add a `forward` method to the distributions for compatibility.
The method is used within prior samplers.