move the kernels out of the algorithms
Kernels are very general function that move one state to another. A
single kernel can be used in different sampling algorithms/schemes with
different adaptive schemes for its parameters.

We thus extract and isolate the kernels. Only Random Walk Metropolis and
Hamiltonian Monte Carlo are currently implemented at the moment, but NUTS
is a likely candidate.

This also makes testing a lot easier.