Fix the batch shape broadcasting
We currently have the parameters directly as arguments of
`broadcast_batch_shape`. This works fine in most cases, but in the
case of the categorical, for instance, the batch shape is a subset of
the parameter's total shape.

We thus modify the function to take as input the shapes that we pass to
it. Fixes issues with the tests on shape for the categorical.