make kernel and sampler for generic
We separate the HMC kernel into several functions: stepper, acceptance
step and a combination of both. This is necessary because we need to
separate the stepping function from the acceptance step in some adaptive
schemes (find reaonable step size and longest batch). This makes the code
mode modular and avoids repetitions.

We also made the sampler more generic so they can be used with virtually
any sampling kernel.