support use of models as distributions
Generative model implicitly define a distribution; for composability
we should be able to use a model as a random variable in another model.
This commit adds this functionality; it is possible to use a model
defined in the same file or imported from another file as a random
variable. This is made possible by parsing the graph of this model and
merging it with that of the model being defined.