pre-compute gradients when initializing verlet integrator
Trajectory integration is where most of the computation time is spent
during inference. I ran a micro-benchmark using the potential of the
harmonic oscillator and found that pre-computing the gradient of the
potential and kinetic energy functions leads to a 10% performance
improvement.

I thus now pre-compute the gradients when initializing the velocity
verlet integrator.