include integrator parameters in functions
The name of the integrator steps are tied to a set of values for their
parameters; furthermore, it makes very little sense to allow users to
change the values of these parameters. We may include later a more
general function so users can play with palindromic iterators, but for
now we stick to the well-tested defaults.