promote momentum shape to variables'
There is currently a tension between the momentum being a 1D array and
the position being a tuple of ND arrays. The two need to reconcile
before we perform operations in the integrators. We considered both the
possibility of flattening the position and promoting the momentum's
shape and it currently makes more sense to promote the momentum's shape.
We modified the code and tests accordingly.

Finally, the log-probability was evaluated in the integrator step
(probably because I did not want to input the logpdf at different
places). However, the evaluation of the logpdf can be costly and
computing it several times for each step is a waste of time and
resources. We now compute it inside the kernel, after the integrator's
update.