refactor the forward sampler
The forward sampler currently makes use of the `sample_shape` of
distributions to draw more than one value. This however leads to many
problems with shapes when sampling prior predictive distributions, that
would involve manipulating the AST further.

However, this does not need to be the case: it is possible to define the
sampler for a single data point and use vmap to batch the computation
for several data points. We draw data points at random with replacement
before batching the computation. This leads to an overall easier and
more robust and still vectorized code.