remove reference to `path_length` in HMC sampler
I believe `path_length` is still in use for historical reasons; however,
it makes more sense to reason in terms of number of integration steps
while it simplifies the cost (no dynamic computation of the number of
integration steps and casting to int). I thus replaced every mention of
`path length` in the HMC proposal and program with
`num_integration_steps`.