transform loglikelihood to potential for HMC
The issue #19 raised the concern that the output of the HMC sampler was
not what would be expected, but instead gave constant values. It turns
out that we fed the loglikelihood to the HMC algorithm instead of the
potential, which is minus the loglikelihood.

We added a function that transforms the loglikelihood into the
potential. Since this is specific to algoritms based on HMC, we added
this function to the HMC program and not the runtimes.