'variables' property returns list consistent with sampler
The sampler returns deterministic and random variables that are ordered
as they appear during a topological traversal of the graph. It is not
the case for "variables" which is used to gather the results of the
sampler, which led to subtle bugs (Issue #19). We changed the behavior
of the model's and graph's 'variables' properties.