change FunctionType type hint to Callable
Mypy does not recognize the intersection between FunctionType and
Callable. Therefore it does not like when `Model` is initialized with a
regular function. This is a shame since it does not recognize
`Callable`'s `__globals__` method.

I thus replace every mention of `FunctionType` by `Callable` and add
a `# type: ignore` in front of the line where we call the `__globals__`
method.