simplify the initialization of the HMC algorithm
To instantiate a HMC program one currently needs to specify the inverse
mass matrix and the square root of the mass matrix. As explained in
issue #18, these are however redundant. We made it such that only the
inverse mass matrix is required; the initialization of the metric takes
care of the computation of the square root of the mass matrix from the
inverse mass matrix.

As a result HMC is easier to use and the code clearer as a byproduct.