add distribution method to broadcast the shape
To keep a simple API when building Bayesian Neural Network we don't want
to have to specify the batching shape of the prior distribution so that
it matches the layer size. Therefore we add a helper function that
re-broadcasts a distribution to a destination shape (here the layer
size) so it can be used in the neural network internals.