improve warmup architecture and speed up
The warmup as it was implemented in warmup.py was very, very slow (a
couple minutes for 100 draws). This is not acceptable. I thus completely
overhauled the design of the warmup in an attempt to speed things up.
The results are promising: using only the "fast stage" of the adaptation
I manage to initialize the step size and perform 1,000 warmup steps in
a little over 10s.

We need to include the "slow stage" before claiming victory, but this is
definitely an encouraging step.