substitute @ for <~ in randvar assignement
By popular acclaim and because it is less confusing for people used to
the matrix multiplication operator, I chose to use the `<~` symbol to
represent random variable assignments in mcx.

`<~` is not a python operator. It is rather the combination of the
comparison and invert operators; a syntactic trick to have a notation
that makes sense while having syntactically correct python. To simplify
the AST parsing we substitute `<~` by the `is` operator using a regexp
beforehand.

remove mentions of @ in docstrings

remove mention of @ in docs

substitute <~ for is operator before parsing

remove mentions of @ in README

change AST parser to support <~

reformat