sample using lax.scan
Having a progress bar is nice, but performance is better. We use
jax.lax.scan to perform the inference instead of using a standard loop +
progress bar. The performance diff is astounding, I am able to run 100k
chains for 10k steps in 30s.

It also appears that the performance bottleneck is in the generation of
initial samples. This can easily be fixed.