run the warmup with lax.scan
Although JIT-compilation dominates the run time of the warmup (can we do
aything about this), executing the loop with `lax.scan` is much faster
than with a progress bar. We thus allow the possibility to run the
warmup without a progress bar but with lax.scan.

There is clearly a design issue here as we need to pass the boolean
`progress_bar` both at the sampler initialization and during a run.
I do like the idea of specifying the `progress_bar` in the run, but
not so much when initializing the sampler. We may need to add a `warmup`
method to the sampler.