s/progress_bar/accelerate
It feels weird to pass `progress_bar` as a keyword argument when it is
the default behavior; furthermore, it implies that the only goal of this
keyword argument is to stop displaying the bar. However the goal is to
accelerate inference using jax.lax.scan; not being able to display the
progress bar is a consequence of using jax.lax.scan. I therefore decided
to rename the keyword argument `progress_bar` into `accelerate` with a
default value of `False`.