give lax.scan and for_loop chains the same structure
lax.scan and the for loop output traces with different shape. We use
`tree_multimap` to stack the values of each step in the chain produced
by the loop. This simplifies greatly the operations to build the chain
thereafter.

Also note that `lax.scan`'s output remains untouched, which means an
even better performance boost for large number of chains and samples.