Integration with ArviZ (#35)
We create a `Trace` object that can be passed to arviZ to compute stats, diagnostics and make some nice plots. To do so we had to harmonize the trace that comes from `lax.scan` and the one that comes from the `for` loop. This leads to a performance issue that will need to be addressed later.