accelerate interactive sampling
interactive sampling suffers from a slowdown due to having to stack the
states after sampling. We solved this issue by raveling states before
they are appended to a list. We then stack the list and unravel the
whole array in a vmap.

We may be able to further accelerate this process by filling a
pre-allocated array, but this is good enough for now.