remove superfluous state updates in Stan warmup
We currently update the state in the main "update" loop as well as in
the two stages' updates. This is not only time consuming during computations
(moving the chain one step) but also during compilation since we need to
pass the kernel factory to the different stages.

We now only move the state in the main loop. The code somehow makes more
sense (only the WarmupState is updated in the two stages) and is faster
to execute.