add prior predictive sampler, improve forward sampler
We added a new `predict` function that can be used for prior as well as
predictive sampling. The prior sampler relies on a compiled function
that only returns the predicted variables.

We improved the forward sampler, the shapes were completely messed up.
Samples now have the shape (broadcasted_input_shape, num_samples). It
happens to simplify the code.

Finally we provided a `forward` method that can come handy when
debugging the model quickly. The user can do `model.forward(x=1.)` to a
get a forward sample. Can be combined with conditioning
`model.do(y=10).forward(x=1.)` for more fun.